requirements.txt
nonebot~=1.9.1
Pillow~=8.3.2
pathlib~=1.0.1
retrying~=1.3.3